# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ERP System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── USERS ───────────────────────────────────────────────────────────────────
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
...
  /attendance/today:
    get:
      operationId: getTodayAttendance
      tags: [attendance]
      summary: Get today's attendance for current user
      responses:
        "200":
          description: Today's record or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
        "404":
          description: No record today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendance/summary:
    get:
      operationId: getAttendanceSummary
      tags: [attendance]
      summary: Monthly attendance summary per employee
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: integer
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Attendance summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceSummary"

  # ─── LEAVE ───────────────────────────────────────────────────────────────────
...
  /complaints:
    get:
      operationId: listComplaints
      tags: [complaints]
      summary: List complaints
      parameters:
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Complaints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Complaint"
    post:
      operationId: createComplaint
      tags: [complaints]
      summary: Submit a complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComplaintBody"
      responses:
        "201":
          description: Complaint submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
  /complaints/{id}:
    get:
      operationId: getComplaint
      tags: [complaints]
      summary: Get complaint by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Complaint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
    patch:
      operationId: updateComplaint
      tags: [complaints]
      summary: Update complaint status or response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComplaintBody"
      responses:
        "200":
          description: Updated complaint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"

  # ─── REQUESTS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ─── AUTH ──────────────────────────────────────────────────────────────────
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        token:
          type: string
      required: [user, token]
    ChangePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required: [currentPassword, newPassword]

    # ─── USERS ─────────────────────────────────────────────────────────────────
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        position:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        employmentStatus:
          type: string
        joinDate:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, email, fullName, role, employmentStatus, createdAt]
...
    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userFullName:
          type: ["string", "null"]
        date:
          type: string
        clockInAt:
          type: ["string", "null"]
        clockOutAt:
          type: ["string", "null"]
        isLate:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, userId, date, isLate, createdAt]
    AttendanceSummary:
      type: object
      properties:
        userId:
          type: integer
        userFullName:
          type: string
        departmentName:
          type: string
        presentDays:
          type: integer
        absentDays:
          type: integer
        lateDays:
          type: integer
        workingDays:
          type: integer
      required: [userId, userFullName, departmentName, presentDays, absentDays, lateDays, workingDays]

    # ─── LEAVE ─────────────────────────────────────────────────────────────────
...
    Complaint:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userFullName:
          type: ["string", "null"]
        departmentId:
          type: integer
        departmentName:
          type: ["string", "null"]
        category:
          type: string
        title:
          type: string
        description:
          type: string
        isAnonymous:
          type: boolean
        status:
          type: string
        assignedTo:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        response:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, departmentId, category, title, description, isAnonymous, status, createdAt, updatedAt]
    CreateComplaintBody:
      type: object
      properties:
        category:
          type: string
        title:
          type: string
        description:
          type: string
        isAnonymous:
          type: boolean
      required: [category, title, description, isAnonymous]
    UpdateComplaintBody:
      type: object
      properties:
        status:
          type: ["string", "null"]
        response:
          type: ["string", "null"]
        assignedTo:
          type: ["integer", "null"]

    # ─── REQUESTS ──────────────────────────────────────────────────────────────
...