# lib/api-zod/src/generated/types/createComplaintBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ERP System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateComplaintBody {
  category: string;
  title: string;
  description: string;
  isAnonymous: boolean;
}